created landing screen and implemented with register screen
