+++ src/main/java/org/app/events/landing/controllers/LandingController.java
@@ -1,0 +1,25 @@
+package org.app.events.landing.controllers;
+
+import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.stage.Stage;
+import javafx.scene.control.Button;
+
+import java.io.IOException;
+
+public class LandingController
+{
+    @FXML
+    Button signupBtn;
+
+    @FXML
+    public void switchToRegister() throws IOException {
+        Stage stage = (Stage)signupBtn.getScene().getWindow();
+        Parent root = FXMLLoader.load(getClass().getClassLoader().getResource("register.fxml"));
+        Scene scene = new Scene(root);
+        stage.setScene(scene);
+        stage.show();
+    }
+}
