--- src/main/java/org/app/events/registration/Main.java
@@ -1,27 +1,0 @@
-package org.app.events.registration;
-
-import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
-import javafx.scene.Parent;
-import javafx.scene.Scene;
-import javafx.stage.Stage;
-import org.app.events.registration.services.UserService;
-
-public class Main extends Application
-{
-
-    @Override
-    public void start(Stage primaryStage) throws Exception
-    {
-        UserService.loadUsersFromFile();
-
-        Parent root = FXMLLoader.load(getClass().getClassLoader().getResource("register.fxml"));
-        primaryStage.setTitle("Registration Example");
-        primaryStage.setScene(new Scene(root, 600, 500));
-        primaryStage.show();
-    }
-
-    public static void main(String[] args) {
-        launch();
-    }
-}
