+++ src/main/java/org/app/events/registration/controllers/RegistrationController.java
@@ -1,12 +1,19 @@
 package org.app.events.registration.controllers;
 
 import javafx.fxml.FXML;
+import javafx.fxml.FXMLLoader;
+import javafx.scene.Parent;
+import javafx.scene.Scene;
+import javafx.scene.control.Button;
 import javafx.scene.control.ChoiceBox;
 import javafx.scene.control.PasswordField;
 import javafx.scene.control.TextField;
 import javafx.scene.text.Text;
+import javafx.stage.Stage;
 import org.app.events.registration.exceptions.UsernameAlreadyExistsException;
 import org.app.events.registration.services.UserService;
+
+import java.io.IOException;
 
 public class RegistrationController {
 
@@ -24,6 +31,8 @@ public class RegistrationController {
     private PasswordField passwordField;
     @FXML
     private ChoiceBox role;
+    @FXML
+    private Button backBtn;
 
     @FXML
     public void initialize()
@@ -48,5 +57,14 @@ public class RegistrationController {
         {
             registrationMessage.setText(e.getMessage());
         }
+    }
+
+    public void changeToLanding() throws IOException
+    {
+        Stage stage = (Stage)backBtn.getScene().getWindow();
+        Parent root = FXMLLoader.load(getClass().getClassLoader().getResource("landing.fxml"));
+        Scene scene = new Scene(root);
+        stage.setScene(scene);
+        stage.show();
     }
 }
