+++ src/main/resources/META-INF/MANIFEST.MF
@@ -1,3 +1,3 @@
 Manifest-Version: 1.0
-Main-Class: org.app.events.registration.Main
+Main-Class: org.app.events.landing.Main
