+++ src/main/resources/landing.fxml
@@ -1,0 +1,19 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<?import javafx.scene.control.*?>
+<?import javafx.scene.layout.*?>
+<?import javafx.geometry.Insets?>
+
+<VBox xmlns="http://javafx.com/javafx"
+      xmlns:fx="http://javafx.com/fxml"
+      fx:controller="org.app.events.landing.controllers.LandingController"
+      prefHeight="400.0" prefWidth="600.0" alignment="center" spacing="10">
+
+    <padding>
+        <Insets bottom="25" left="25" right="25" top="25"/>
+    </padding>
+
+    <Label text="Welcome to the Events App!" />
+    <Button id="loginBtn" text="Log In"/>
+    <Button fx:id="signupBtn" onAction="#switchToRegister" text="Sign up"/>
+
+</VBox>
