+++ src/main/resources/register.fxml
@@ -16,7 +16,7 @@
             <Insets bottom="25" left="25" right="25" top="25"/>
         </padding>
 
-        <Text id="title-text" text="SEF - Events Application" GridPane.columnIndex="0" GridPane.columnSpan="2" GridPane.rowIndex="0"/>
+        <Label id="title-text" text="SEF - Events Application" GridPane.columnIndex="0" GridPane.columnSpan="2" GridPane.rowIndex="0"/>
 
         <Label text="Username:" GridPane.columnIndex="0" GridPane.rowIndex="1"/>
         <TextField fx:id="usernameField" GridPane.columnIndex="1" GridPane.rowIndex="1"/>
@@ -39,7 +39,7 @@
 
         <Text fx:id="registrationMessage" GridPane.columnIndex="1" GridPane.rowIndex="10"/>
         <HBox spacing="10" GridPane.rowIndex="8" GridPane.columnIndex="1" alignment="bottom_right">
-                <Button id="backBtn" text="Back"/>
+                <Button fx:id="backBtn" onAction="#changeToLanding" text="Back"/>
                 <Button id="registerBtn" onAction="#handleRegisterAction" text="Sign up" />
         </HBox>
